prototypes: fix penetrations using spring forces.
The spring constant seems to be insanely large, but nothing less is
reliably working. Also reordered and slightly documented the impulse
expression, with spring forces it could happen that the impulse could be
computed in wrong direction (pushing the bodies towards each other
instead of pushing them apart), fixed that.